fix full battles
